+++ docs/docs/auto-docs/App/functions/default.md
@@ -10,7 +10,7 @@ Defined in: [src/App.tsx:109](https://github.com/PalisadoesFoundation/talawa-adm
 
 This is the main function for our application. It sets up all the routes and components,
 defining how the user can navigate through the app. The function uses React Router's `Routes`
-and `Route`  components to map different URL paths to corresponding screens and components.
+and `Route` components to map different URL paths to corresponding screens and components.
 
 ## Important Details
 - **UseEffect Hook**: This hook checks user authentication status using the `CHECK_AUTH` GraphQL query.
