--- docs/docs/auto-docs/components/OrganizationCard/OrganizationCard/functions/default.md
@@ -1,29 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: default()
-
-> **default**(`__namedParameters`): `JSX.Element`
-
-Defined in: [src/components/OrganizationCard/OrganizationCard.tsx:46](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrganizationCard/OrganizationCard.tsx#L46)
-
-Component to display an organization's card with its image and owner details.
-Displays an organization card with options to join or manage membership.
-
-Shows the organization's name, image, description, address, number of admins and members,
-and provides buttons for joining, withdrawing membership requests, or visiting the organization page.
-
-## Parameters
-
-### \_\_namedParameters
-
-[`InterfaceOrganizationCardProps`](../../../../types/Organization/interface/interfaces/InterfaceOrganizationCardProps.md)
-
-## Returns
-
-`JSX.Element`
-
-JSX element representing the organization card.
-
-The organization card component.
