+++ docs/docs/auto-docs/components/UserPortal/OrganizationCard/OrganizationCard/functions/default.md
@@ -4,16 +4,26 @@
 
 # Function: default()
 
-> **default**(`props`): `JSX.Element`
+> **default**(`__namedParameters`): `JSX.Element`
 
-Defined in: [src/components/UserPortal/OrganizationCard/OrganizationCard.tsx:44](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/OrganizationCard/OrganizationCard.tsx#L44)
+Defined in: [src/components/UserPortal/OrganizationCard/OrganizationCard.tsx:47](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/OrganizationCard/OrganizationCard.tsx#L47)
+
+Component to display an organization's card with its image and owner details.
+Displays an organization card with options to join or manage membership.
+
+Shows the organization's name, image, description, address, number of admins and members,
+and provides buttons for joining, withdrawing membership requests, or visiting the organization page.
 
 ## Parameters
 
-### props
+### \_\_namedParameters
 
 [`InterfaceOrganizationCardProps`](../../../../../types/Organization/interface/interfaces/InterfaceOrganizationCardProps.md)
 
 ## Returns
 
 `JSX.Element`
+
+JSX element representing the organization card.
+
+The organization card component.
