+++ src/components/UserPortal/OrganizationCard/OrganizationCard.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import { useEffect, useState } from 'react';
 import styles from '../../../style/app-fixed.module.css';
 import { Button } from 'react-bootstrap';
 import { Tooltip } from '@mui/material';
@@ -14,12 +14,14 @@ import {
   ORGANIZATION_LIST,
   USER_JOINED_ORGANIZATIONS_PG,
 } from 'GraphQl/Queries/Queries';
-import { getItem } from 'utils/useLocalstorage';
 import Avatar from 'components/Avatar/Avatar';
 import { useNavigate } from 'react-router-dom';
+import type { ApolloError } from '@apollo/client';
 import type { InterfaceOrganizationCardProps } from 'types/Organization/interface';
+import { getItem } from 'utils/useLocalstorage';
 
 /**
+ * Component to display an organization's card with its image and owner details.
  * Displays an organization card with options to join or manage membership.
  *
  * Shows the organization's name, image, description, address, number of admins and members,
@@ -37,11 +39,36 @@ import type { InterfaceOrganizationCardProps } from 'types/Organization/interfac
  * @param userRegistrationRequired - Indicates if user registration is required to join the organization.
  * @param membershipRequests - The list of membership requests with user IDs.
  *
+ * @param props - Properties for the organization card.
+ * @returns JSX element representing the organization card.
  * @returns The organization card component.
  */
-const userId: string | null = getItem('Talawa-admin', 'userId');
 
-function organizationCard(props: InterfaceOrganizationCardProps): JSX.Element {
+function OrganizationCard({
+  id,
+  name,
+  image,
+  description,
+  admins,
+  members,
+  address,
+  membershipRequestStatus,
+  userRegistrationRequired,
+  membershipRequests,
+  isJoined,
+}: InterfaceOrganizationCardProps): JSX.Element {
+  const [userId, setUserId] = useState<string | null>(null);
+
+  useEffect(() => {
+    try {
+      const storedUserId = getItem('Talawa-admin', 'id');
+      setUserId(storedUserId as string | null);
+    } catch (error) {
+      // Handle localStorage error silently
+      console.error('Failed to get userId from localStorage:', error);
+    }
+  }, []);
+
   const { t } = useTranslation('translation', {
     keyPrefix: 'users',
   });
@@ -50,13 +77,13 @@ function organizationCard(props: InterfaceOrganizationCardProps): JSX.Element {
 
   // Mutations for handling organization memberships
   const [sendMembershipRequest] = useMutation(SEND_MEMBERSHIP_REQUEST, {
-    refetchQueries: [{ query: ORGANIZATION_LIST }],
+    refetchQueries: [{ query: ORGANIZATION_LIST, variables: { id } }],
   });
   const [joinPublicOrganization] = useMutation(JOIN_PUBLIC_ORGANIZATION, {
-    refetchQueries: [{ query: ORGANIZATION_LIST }],
+    refetchQueries: [{ query: ORGANIZATION_LIST, variables: { id } }],
   });
   const [cancelMembershipRequest] = useMutation(CANCEL_MEMBERSHIP_REQUEST, {
-    refetchQueries: [{ query: ORGANIZATION_LIST }],
+    refetchQueries: [{ query: ORGANIZATION_LIST, variables: { id } }],
   });
   const { refetch } = useQuery(USER_JOINED_ORGANIZATIONS_PG, {
     variables: { id: userId, first: 5 },
@@ -69,17 +96,19 @@ function organizationCard(props: InterfaceOrganizationCardProps): JSX.Element {
    */
   async function joinOrganization(): Promise<void> {
     try {
-      if (props.userRegistrationRequired) {
+      if (userRegistrationRequired) {
         await sendMembershipRequest({
           variables: {
-            organizationId: props.id,
+            organizationId: id,
           },
         });
         toast.success(t('MembershipRequestSent') as string);
       } else {
         await joinPublicOrganization({
           variables: {
-            organizationId: props.id,
+            input: {
+              organizationId: id,
+            },
           },
         });
         toast.success(t('orgJoined') as string);
@@ -87,7 +116,9 @@ function organizationCard(props: InterfaceOrganizationCardProps): JSX.Element {
       refetch();
     } catch (error: unknown) {
       if (error instanceof Error) {
-        if (error.message === 'User is already a member') {
+        const apolloError = error as ApolloError;
+        const errorCode = apolloError.graphQLErrors?.[0]?.extensions?.code;
+        if (errorCode === 'ALREADY_MEMBER') {
           toast.error(t('AlreadyJoined') as string);
         } else {
           toast.error(t('errorOccured') as string);
@@ -100,96 +131,113 @@ function organizationCard(props: InterfaceOrganizationCardProps): JSX.Element {
    * Handles withdrawing a membership request. Finds the request for the current user and cancels it.
    */
   async function withdrawMembershipRequest(): Promise<void> {
-    const membershipRequest = props.membershipRequests.find(
-      (request) => request.user._id === userId,
+    if (!userId) {
+      toast.error(t('UserIdNotFound') as string);
+      return;
+    }
+
+    const membershipRequest = membershipRequests.find(
+      (request) => request.user.id === userId,
     );
 
-    await cancelMembershipRequest({
-      variables: {
-        membershipRequestId: membershipRequest?._id,
-      },
-    });
+    try {
+      if (!membershipRequest) {
+        toast.error(t('MembershipRequestNotFound') as string);
+        return;
+      }
+
+      await cancelMembershipRequest({
+        variables: {
+          membershipRequestId: membershipRequest.id,
+        },
+      });
+
+      toast.success(t('MembershipRequestWithdrawn') as string);
+    } catch (error: unknown) {
+      if (process.env.NODE_ENV === 'development') {
+        console.error('Failed to withdraw membership request:', error);
+      }
+      toast.error(t('errorOccured') as string);
+    }
   }
 
   return (
-    <>
-      <div className={styles.orgCard}>
-        <div className={styles.innerContainer}>
-          <div className={styles.orgImgContainer}>
-            {props.image ? (
-              <img src={props.image} alt={`${props.name} image`} />
-            ) : (
-              <Avatar
-                name={props.name}
-                alt={`${props.name} image`}
-                dataTestId="emptyContainerForImage"
-              />
-            )}
-          </div>
-          <div className={styles.content}>
-            <Tooltip title={props.name} placement="top-end">
-              <h4 className={`${styles.orgName} fw-semibold`}>{props.name}</h4>
-            </Tooltip>
-            <h6 className={`${styles.orgdesc} fw-semibold`}>
-              <span>{props.description}</span>
-            </h6>
-            {props.address && props.address.city && (
-              <div className={styles.address}>
-                <h6 className="text-secondary">
-                  <span className="address-line">{props.address.line1}, </span>
-                  <span className="address-line">{props.address.city}, </span>
-                  <span className="address-line">
-                    {props.address.countryCode}
-                  </span>
-                </h6>
-              </div>
-            )}
-            <h6 className={styles.orgadmin}>
-              <div>
-                {tCommon('admins')}: <span>{props.admins?.length}</span>
-              </div>
-              <div>
-                {tCommon('members')}: <span>{props.members?.length}</span>
-              </div>
-            </h6>
-          </div>
+    <div className={styles.orgCard}>
+      <div className={styles.innerContainer}>
+        <div className={styles.orgImgContainer}>
+          {image ? (
+            <img src={image} alt={`${name} image`} />
+          ) : (
+            <Avatar
+              name={name}
+              alt={`${name} image`}
+              dataTestId="emptyContainerForImage"
+            />
+          )}
         </div>
-        {props.membershipRequestStatus === 'accepted' && (
-          <Button
-            data-testid="manageBtn"
-            className={styles.addButton}
-            onClick={() => {
-              navigate(`/user/organization/${props.id}`);
-            }}
-            style={{ width: '8rem' }}
-          >
-            {t('visit')}
-          </Button>
-        )}
-
-        {props.membershipRequestStatus === 'pending' && (
-          <Button
-            variant="danger"
-            onClick={withdrawMembershipRequest}
-            data-testid="withdrawBtn"
-            className={styles.withdrawBtn}
-          >
-            {t('withdraw')}
-          </Button>
-        )}
-        {props.membershipRequestStatus === '' && (
-          <Button
-            onClick={joinOrganization}
-            data-testid="joinBtn"
-            className={styles.outlineBtn}
-            style={{ width: '8rem' }}
-          >
-            {t('joinNow')}
-          </Button>
-        )}
+        <div className={styles.content}>
+          <Tooltip title={name} placement="top-end">
+            <h4 className={`${styles.orgName} fw-semibold`}>{name}</h4>
+          </Tooltip>
+          <h6 className={`${styles.orgdesc} fw-semibold`}>
+            <span>{description}</span>
+          </h6>
+          {address && address.city && (
+            <div className={styles.address}>
+              <h6 className="text-secondary">
+                <span className="address-line">{address.line1}, </span>
+                <span className="address-line">{address.city}, </span>
+                <span className="address-line">{address.countryCode}</span>
+              </h6>
+            </div>
+          )}
+          <h6 className={styles.orgadmin}>
+            <div>
+              {tCommon('admins')}: <span>{admins?.length}</span>
+            </div>
+            <div>
+              {tCommon('members')}: <span>{members?.length}</span>
+            </div>
+          </h6>
+        </div>
       </div>
-    </>
+
+      {isJoined && (
+        <Button
+          data-testid="manageBtn"
+          className={styles.addButton}
+          onClick={() => {
+            navigate(`/user/organization/${id}`);
+          }}
+          style={{ width: '8rem' }}
+        >
+          {t('visit')}
+        </Button>
+      )}
+
+      {membershipRequestStatus === 'pending' && !isJoined && (
+        <Button
+          variant="danger"
+          onClick={withdrawMembershipRequest}
+          data-testid="withdrawBtn"
+          className={styles.withdrawBtn}
+        >
+          {t('withdraw')}
+        </Button>
+      )}
+
+      {membershipRequestStatus === '' && !isJoined && (
+        <Button
+          onClick={joinOrganization}
+          data-testid="joinBtn"
+          className={styles.outlineBtn}
+          style={{ width: '8rem' }}
+        >
+          {t('joinNow')}
+        </Button>
+      )}
+    </div>
   );
 }
 
-export default organizationCard;
+export default OrganizationCard;
